ml-services/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import numpy as np
import pandas as pd
from datetime import datetime, timedelta
import os
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)
CORS(app)

# Mock ML models (in production, these would be trained models)
class WorkflowOptimizer:
    def __init__(self):
        pass
    
    def detect_bottlenecks(self, workflow_data):
        """Detect bottlenecks in workflow"""
        # Mock implementation
        bottlenecks = []
        if workflow_data.get('avg_cycle_time', 0) > 7:
            bottlenecks.append({
                'stage': 'review',
                'issue': 'High cycle time',
                'severity': 'high',
                'suggestion': 'Consider parallelizing review process'
            })
        return bottlenecks
    
    def optimize_workflow(self, workflow_id):
        """Generate workflow optimization suggestions"""
        return {
            'bottlenecks': self.detect_bottlenecks({}),
            'suggestions': [
                {
                    'type': 'parallelize',
                    'stage': 'review',
                    'impact': 'high',
                    'description': 'Parallelize review stage to reduce cycle time'
                }
            ],
            'efficiency_score': 0.75
        }

class ForecastModel:
    def __init__(self):
        pass
    
    def forecast_completion(self, workflow_data):
        """Forecast workflow completion using Monte Carlo simulation"""
        # Mock Monte Carlo simulation
        avg_velocity = workflow_data.get('avg_velocity', 5)  # items per day
        remaining_items = workflow_data.get('remaining_items', 20)
        
        # Simulate 1000 scenarios
        simulations = []
        for _ in range(1000):
            # Add some randomness
            velocity = np.random.normal(avg_velocity, avg_velocity * 0.2)
            days = max(1, remaining_items / velocity)
            simulations.append(days)
        
        mean_days = np.mean(simulations)
        std_days = np.std(simulations)
        
        completion_date = datetime.now() + timedelta(days=int(mean_days))
        lower_bound = datetime.now() + timedelta(days=int(mean_days - std_days))
        upper_bound = datetime.now() + timedelta(days=int(mean_days + std_days))
        
        return {
            'completion_date': completion_date.isoformat(),
            'confidence': 0.75,
            'confidence_interval': {
                'lower': lower_bound.isoformat(),
                'upper': upper_bound.isoformat()
            }
        }

optimizer = WorkflowOptimizer()
forecast_model = ForecastModel()

@app.route('/health', methods=['GET'])
def health():
    return jsonify({'status': 'ok'})

@app.route('/api/workflows/<workflow_id>/optimize', methods=['GET'])
def optimize_workflow(workflow_id):
    """Get workflow optimization suggestions"""
    try:
        result = optimizer.optimize_workflow(workflow_id)
        return jsonify(result)
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/analytics/forecast/<workflow_id>', methods=['GET'])
def forecast_workflow(workflow_id):
    """Forecast workflow completion"""
    try:
        # Mock workflow data - in production, fetch from database
        workflow_data = {
            'avg_velocity': 5,
            'remaining_items': 20
        }
        result = forecast_model.forecast_completion(workflow_data)
        return jsonify(result)
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/analytics/bottlenecks/<workflow_id>', methods=['GET'])
def get_bottlenecks(workflow_id):
    """Get workflow bottlenecks"""
    try:
        bottlenecks = optimizer.detect_bottlenecks({})
        return jsonify(bottlenecks)
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=True)

