ml-services/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.0

